💅 Mostly a reorg of files to a screaming architecture
The screaming architecture is one which 'screams' what its purpose
is. In this case, I'm reorganizing my project along business logic
lines in the hopes that I'm clearer in what is going on. This also
means that related code is much closer together "physically".

The term "screaming architecture" comes from Uncle Bob circa 2011:
https://blog.cleancoder.com/uncle-bob/2011/09/30/Screaming-Architecture.html