🧪 Fix tests to account for authorization
Now that we have the middleware in place, we need to set the
authorization header on each of the network requests that are
protected.

This is done using the `set` method from supertest.

The other big change is that the responses are now all json objects
with a data or an error key on the body.